# swagger/merged_workers.yaml
openapi: 3.0.0
info:
  title: Workday Workers API (RAG-Optimized)
  version: 1.0.0
  description: Consolidated read-only endpoints for the Workers module, curated for semantic routing and RAG search.
paths:
  /workers:
    get:
      summary: Retrieves a collection of workers.
      description: |
        Use this endpoint to search or list all workers in Workday. 
        Note: This is NOT for retrieving a specific worker's profile, pay, time off, organization, or history (those require a worker ID and separate endpoints).
      operationId: retrieves_a_collection_of_workers
      x-module: workers
      x-action: read
      x-scope: collection
      x-intent-triggers:
        - "list of all workers"
        - "give me the list of workers"
        - "list all workers in Workday"
        - "list all employees"
        - "search for a worker"
        - "find employee by name"
        - "get all staff"
        - "show workers"
        - "fetch employee list"
        - "search people"
        - "who works here"
        - "show terminated employees"
      parameters:
        - name: search
          in: query
          required: false
          description: Filter workers by name or search term.
          schema:
            type: string
        - name: includeTerminatedWorkers
          in: query
          required: false
          description: Flag to include terminated employees in the list.
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          description: Maximum number of worker records to return.
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          description: Number of records to skip for pagination.
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved a collection of workers.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          pattern: '^(([0-9a-f]{32})|(\S+=\S+))$'
                        descriptor:
                          type: string
                        href:
                          type: string
                        businessTitle:
                          type: string
                        primaryWorkEmail:
                          type: string
                        isManager:
                          type: boolean
                        primaryWorkPhone:
                          type: string
                        primarySupervisoryOrganization:
                          type: object
                          required:
                            - id
                          properties:
                            id:
                              type: string
                              pattern: '^(([0-9a-f]{32})|(\S+=\S+))$'
                            descriptor:
                              type: string
                            href:
                              type: string
                  total:
                    type: integer
        '400':
          description: "Invalid request. (https://community.workday.com/rest/error-messages)"
        '401':
          description: "Unauthorized: Missing or invalid credentials. (https://community.workday.com/rest/error-messages)"
        '403':
          description: "User has insufficient permissions. (https://community.workday.com/rest/error-messages)"
        '404':
          description: "Resource not found. (https://community.workday.com/rest/error-messages)"
  /workers/{ID}:
    get:
      summary: Retrieves a worker instance.
      description: |
        Use this endpoint to look up a specific person or retrieve details for a single worker using their Workday unique identifier or Employee ID. Use 'me' as the ID to perform a self-lookup of the current logged-in user's profile.
        Note: This is NOT for retrieving specific sub-resources like pay slips, time off plans, time off entries, organization structures, history logs, direct reports, or inbox tasks. Those require separate sub-resource endpoints.
      operationId: retrieves_a_worker_instance
      x-module: workers
      x-action: read
      x-scope: single
      x-intent-triggers:
        - "get employee details having employee ID : \"1234\""
        - "fetch worker profile who's emp ID is 1234"
        - "who is this employee with emp 1234"
        - "show me worker info having employee id 1234"
        - "get my own profile"
        - "lookup a specific person"
        - "retrieve worker by ID"
      parameters:
        - name: ID
          in: path
          required: true
          description: "The unique identifier of the worker or 'me' for a self-lookup."
          schema:
            type: string
            pattern: '^(([0-9a-f]{32})|(\S+=\S+)|me)$'
      responses:
        '200':
          description: Successfully retrieved the specific worker instance profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  descriptor:
                    type: string
                  href:
                    type: string
                  businessTitle:
                    type: string
                  isManager:
                    type: boolean
                  primaryWorkEmail:
                    type: string
                  primaryWorkPhone:
                    type: string
                  primaryWorkAddressText:
                    type: string
                  dateOfBirth:
                    type: string
                    format: date-time
                  yearsOfService:
                    type: number
                    format: double
                  supervisoryOrganizationsManaged:
                    type: string
                    format: uri
                  primarySupervisoryOrganization:
                    type: object
                    properties:
                      id:
                        type: string
                      descriptor:
                        type: string
                      href:
                        type: string
                  location:
                    type: object
                    properties:
                      id:
                        type: string
                      descriptor:
                        type: string
                      href:
                        type: string
        '400':
          description: "Invalid request. (https://community.workday.com/rest/error-messages)"
        '401':
          description: "Unauthorized: Missing or invalid credentials. (https://community.workday.com/rest/error-messages)"
        '403':
          description: "User has insufficient permissions. (https://community.workday.com/rest/error-messages)"
        '404':
          description: "Resource not found. (https://community.workday.com/rest/error-messages)"
...